Distance function now depends on the time. This has the following consequences: - We can have animations (yay!) - Managing the render cache of the painters is harder (nooo)
There is now a trade off between the actual and felt framerate. If the framerate is too high, the painter has only time to update a few pixels at each frame. This means that only a small part of the image will be updated, and since the part is random, it might be a boring time and the image will not update very fast. The frame rate will then feel low.

OTOH with the current frame rate of 10fps (which sounds low), the painters have a lot of time to update a bug chunk of the image between two frames and so it visually feels like the image is updating fast. Of course the image is in absolute terms very slow at updating.